techpack/camera: Eliminate floating-point datatypes

+++ techpack/camera/drivers/oplus/cam_sensor_module/cam_ois/onsemi_fw/LC898124/E2PromCmd.c
@@ -1005,7 +1005,7 @@ UINT8	WrHallLnData(  UINT8 UcMode, mlLinearityValue *linval )
 	UINT32 UlReadVal, UlCnt;
 	UINT8 ans, data[CHECK_SUM_NUM], cnt, UcReadVal[2];
 	UINT32 ReadVerify, Parity;
-	double		*pPosX, *pPosY;
+	long		*pPosX, *pPosY;
 	UINT32	PosDifX, PosDifY;
 	DSPVER Info;
